+++ database/nostr-lmdb/src/store/error.rs
@@ -28,6 +28,8 @@ pub enum Error {
     WrongEventKind,
     /// Not found
     NotFound,
+    /// Batched transaction failed - sent to operations that didn't cause the error
+    BatchTransactionFailed,
 }
 
 impl std::error::Error for Error {}
@@ -45,6 +47,7 @@ impl fmt::Display for Error {
             Self::MpscSend => write!(f, "mpsc channel send error"),
             Self::NotFound => write!(f, "Not found"),
             Self::WrongEventKind => write!(f, "Wrong event kind"),
+            Self::BatchTransactionFailed => write!(f, "Batched transaction failed"),
         }
     }
 }